Start implementing MVVM pattern for Book Details: BookViewModel is backed by two types of Models - GoogleBookModel and FirebaseBookModel, depending on the datasource.
GoogleBookModel and FirebaseBookModel are not mutual exclusive because we store state (# pages read, status of the book, rating etc) for a book from Google API in our Firebase Realtime Database instance.

We create FirebaseBookModel as soon as the user interacts with the GoogleBookModel. All the logic mentioned is encapsulated in BookViewModel.